Refactor: Extract providers and services into separate files
